# plugins/datasource-common/src/schema/openapi/generated/models/CreateDatasource.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************
import { ExistsIn } from '../models/ExistsIn.model';
import { RhDataClassifications } from '../models/RhDataClassifications.model';

/**
 * @public
 */
export interface CreateDatasource {
  name: string;
  title: string;
  namespace: string;
  description?: string;
  aiRelated: CreateDatasourceAiRelatedEnum;
  owner: string;
  steward: string;
  type: string;
  usage: string;
  location: string;
  classification: RhDataClassifications;
  existsIn: Array<ExistsIn>;
  system?: string;
  dependencyOf?: Array<string>;
  dependsOn?: Array<string>;
}

/**
 * @public
 */
export type CreateDatasourceAiRelatedEnum = 'true' | 'false';