# plugins/datasource-common/src/schema/openapi/generated/models/Datasource.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************
import { ExistsIn } from '../models/ExistsIn.model';
import { RhDataClassifications } from '../models/RhDataClassifications.model';

/**
 * @public
 */
export interface Datasource {
  name: string;
  title: string;
  namespace: string;
  description?: string;
  aiRelated: DatasourceAiRelatedEnum;
  owner: string;
  steward: string;
  type: string;
  usage: string;
  location: string;
  classification: RhDataClassifications;
  existsIn: Array<ExistsIn>;
  system?: string;
  dependencyOf?: Array<string>;
  dependsOn?: Array<string>;
  id: string;
  createdAt: string;
  createdBy: string;
  updatedAt: string;
  updatedBy: string;
}

/**
 * @public
 */
export type DatasourceAiRelatedEnum = 'true' | 'false';