# plugins/datasource-common/src/schema/openapi/generated/models/ExistsIn.model.ts
//

// ******************************************************************
// * THIS IS AN AUTOGENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. *
// ******************************************************************
import { DataStoreNames } from '../models/DataStoreNames.model';

/**
 * @public
 */
export interface ExistsIn {
  name: DataStoreNames;
  description?: string;
}